TSK-521: activated query tasks by workbasket tasks and moved one tests to consolidate all corresponding tests in one class.
