+++ lib/taskana-core/src/test/java/acceptance/task/QueryTasksByWorkbasketAccTest.java
@@ -1,12 +1,19 @@
 package acceptance.task;
 
-import org.junit.Ignore;
+import static org.hamcrest.core.IsEqual.equalTo;
+import static org.junit.Assert.assertThat;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import acceptance.AbstractAccTest;
 import pro.taskana.KeyDomain;
 import pro.taskana.TaskService;
+import pro.taskana.TaskSummary;
 import pro.taskana.exceptions.NotAuthorizedToQueryWorkbasketException;
 import pro.taskana.security.JAASRunner;
 import pro.taskana.security.WithAccessId;
@@ -21,7 +28,31 @@ public class QueryTasksByWorkbasketAccTest extends AbstractAccTest {
         super();
     }
 
-    @Ignore  // BB
+    @WithAccessId(
+        userName = "teamlead_1",
+        groupNames = {"group_1"})
+    @Test
+    public void testQueryForWorkbasketKeyDomain() {
+        TaskService taskService = taskanaEngine.getTaskService();
+        List<KeyDomain> workbasketIdentifiers = Arrays.asList(new KeyDomain("GPK_KSC", "DOMAIN_A"),
+            new KeyDomain("USER_1_2", "DOMAIN_A"));
+
+        List<TaskSummary> results = taskService.createTaskQuery()
+            .workbasketKeyDomainIn(workbasketIdentifiers.toArray(new KeyDomain[0]))
+            .list();
+        assertThat(results.size(), equalTo(42));
+
+        String[] ids = results.stream()
+            .map(t -> t.getWorkbasketSummary().getId())
+            .collect(Collectors.toList())
+            .toArray(new String[0]);
+
+        List<TaskSummary> result2 = taskService.createTaskQuery()
+            .workbasketIdIn(ids)
+            .list();
+        assertThat(result2.size(), equalTo(42));
+    }
+
     @WithAccessId(
         userName = "user_1_1",
         groupNames = {"group_1"})
@@ -33,7 +64,6 @@ public class QueryTasksByWorkbasketAccTest extends AbstractAccTest {
             .list();
     }
 
-    @Ignore  // BB
     @WithAccessId(
         userName = "user_1_1",
         groupNames = {"group_1"})
